+++ rs/src/complete/tag.rs
@@ -1435,7 +1435,7 @@ mod tests {
 
   //  #[test]
   //  fn test_fuzzing() {
-  //    let artifact: &[u8] = include_bytes!("../../fuzz/artifacts/tag/crash-7176426d9af74c41413d9fb11b2d622c0d0342f0");
+  //    let artifact: &[u8] = include_bytes!("../../fuzz/artifacts/tag/crash-4010bd249f7cb7cc86fa23e9060ce750cc1ff8ee");
   //    let (swf_version, input_bytes) = artifact.split_first().unwrap();
   //    let _ = parse_tag(input_bytes, *swf_version);
   //  }